+++ cupy/_functional/piecewise.py
