--- cupy/functional/__init__.py
@@ -1,2 +1,0 @@
-# "NOQA" to suppress flake8 warning
-from cupy.functional import piecewise  # NOQA
